Move kmeans_clusters to core/base because it will be also used in schemeshard
