+++ ydb/core/tx/datashard/build_index/kmeans_helper.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "common_helper.h"
-#include "kmeans_clusters.h"
+#include <ydb/core/base/kmeans_clusters.h>
 #include <ydb/core/base/table_index.h>
 #include <ydb/core/tx/datashard/buffer_data.h>
 #include <ydb/core/tx/datashard/datashard_user_table.h>
