+++ ydb/core/tx/datashard/ya.make
@@ -215,7 +215,6 @@ SRCS(
     wait_for_stream_clearance_unit.cpp
 
     build_index/prefix_kmeans.cpp
-    build_index/kmeans_clusters.cpp
     build_index/kmeans_helper.cpp
     build_index/local_kmeans.cpp
     build_index/sample_k.cpp
